Refactor: 共通のエラーレスポンス処理のディレクトリを親の階層に移動
