+++ internal/controller/error/response/gen/type.gen.go
