+++ internal/controller/error/response/http_error_test.go
